update tests after fresnel fix
