+++ tests/conftest.py
@@ -4,22 +4,22 @@ import pytest
 
 @pytest.fixture
 def get_matlab_data():
-    return pd.read_csv("./tests/matlab_benchmark_data.csv", header=None)
+    return pd.read_csv("./tests/test_data/matlab_benchmark_data.csv", header=None)
 
 
 @pytest.fixture
 def get_matlab_data_toon():
-    return pd.read_csv("./tests/matlab_benchmark_data_toon.csv", header=None)
+    return pd.read_csv("./tests/test_data/matlab_benchmark_data_toon.csv", header=None)
 
 
 @pytest.fixture
 def get_python_data():
-    return pd.read_csv("./tests/py_benchmark_data.csv", header=None).transpose()
+    return pd.read_csv("./tests/test_data/py_benchmark_data.csv", header=None).transpose()
 
 
 @pytest.fixture
 def get_python_data_toon():
-    return pd.read_csv("./tests/py_benchmark_data_toon.csv", header=None).transpose()
+    return pd.read_csv("./tests/test_data/py_benchmark_data_toon.csv", header=None).transpose()
 
 
 @pytest.fixture
